pyproject.toml
[build-system]
requires = ["hatchling<=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "awesom"
version = "0.1.0"
description = "Site-of-metabolism prediction with aleatoric and epistemic uncertainty quantification."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT License"}
authors = [
  {name = "Roxane Jacob", email = "roxane.axel.jacob@univie.ac.at"},
]

dependencies = [
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "numpy>=1.26.4",
    "optuna>=4.1.0",
    "pandas>=1.5.2,<2.2.0",
    "rdkit==2023.3.3",
    "scikit-learn>=1.6.1",
    "tensorboard>=2.18.0",
    "torch>=2.5.1",
    "torchmetrics>=1.0.3",
    "torch-geometric>=2.6.1",
    "PyYAML>=6.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
  "black",
  "isort",
  "pylint",
  "pydocstyle",
]
vis = [
  "ipykernel", 
  "ipywidgets"
]

[tool.isort]
profile = "black"
